Move TotalTimeRootMetaMixin to general timed serializers
This mixin might be used by any end point calculation
total of durations.
