+++ timed/serializers.py
@@ -1,3 +1,7 @@
+from datetime import timedelta
+
+from django.db.models import Sum
+from django.utils.duration import duration_string
 from rest_framework_json_api.serializers import Serializer
 
 
@@ -36,3 +40,19 @@ class DictObjectSerializer(Serializer):
         else:
             instance = [DictObject(**entry) for entry in instance]
         return super().__new__(cls, instance, **kwargs)
+
+
+class TotalTimeRootMetaMixin(object):
+    duration_field = 'duration'
+
+    def get_root_meta(self, resource, many):
+        """Add total hours over whole result (not just page) to meta."""
+        if many:
+            view = self.context['view']
+            queryset = view.filter_queryset(view.get_queryset())
+            data = queryset.aggregate(total_time=Sum(self.duration_field))
+            data['total_time'] = duration_string(
+                data['total_time'] or timedelta(0)
+            )
+            return data
+        return {}
